+++ src/dialog/DonationDialog.test.jsx
@@ -1,11 +1,20 @@
+import MaterialIconButton from '@material-ui/core/IconButton'
 import React from 'react'
 import { shallow, mount } from 'enzyme'
 
+import Button from '../Button'
 import DonationDialog from './DonationDialog'
 
-import MaterialIconButton from '@material-ui/core/IconButton'
-
 describe('<DonationDialog />', () => {
+  describe('onClick callback', () => {
+    it('is called when button is clicked', () => {
+      const onClick = jest.fn()
+      const wrapper = shallow(<DonationDialog open onClick={onClick} />)
+      wrapper.dive().find(Button).simulate('click')
+      expect(onClick).toHaveBeenCalled()
+    })
+  })
+
   describe('onClose callback', () => {
     it('is called when dialog is closed', () => {
       const onClose = jest.fn()