+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Change the `DonationDialog` `href` prop to an `onClick` callback
+
 ## 1.23.2 (2019-04-24)
 
 * Add `!important` to `DonationDialog`, `DonationBanner` and `ThinDonationBanner` button colours